Make pasting into the REPL much simpler
Previously, the REPL would try to determine if you had pasted a complete expression, and, if that expression was followed by a newline, it would evaluate it.  If there were multiple expressions, it would evaluate each one by one. However, this often fails for me when pasting incomplete snippets that I intend to edit in the command line, and since we have such good multi-line editing support I don't even think it's necessary.

This patch simply removes those smarts.  All pasted text is entered into the buffer verbatim (except for an unindentation pass if the cursor is at the beginning of the buffer). The pasted text is not sent for evaluation, even if followed by trailing newlines.  Additionally, this fixes #6664 as the cursor ends up at the end of the pasted snippet.