Do not use setValueName to name metadata nodes
Naming metadata nodes using setValueName() causes references to the
nodes to be printed as %name, while the actual metadata definitions keep
using a numerical reference in form of !19. This mismatch prevents the
LLVM-IR reader to parse the created LLVM-IR. We avoid the creation of
invalid LLVM-IR by only providing the name of metadata nodes in the
createTBAANode() call, but by not using setValueName to name the nodes.