README: Remove old terminal binding instructions
There were previously two terminal binding tips: Bind backspace to `^H` and shift-enter to `\n` instead of `\r`.

For backspace, the new REPL also binds ASCII 127 to backspace. If there are terminals sending different keycodes than either 127 or `^H` (which is equivalent to `\b`) then they are either configured totally wrong or we should attempt to bind those keycodes, too.

For return, there is now no difference between `\r` and `\n`, so there is no point in directing people to make that change.  Line feeds are entered with meta-enter.