Deprecate bitwise &, | whose parsing changes…
with a change in precedence to be lower than comparisons.

Due to the way the parser works, it is much easier to flag the locations where the bitwise operators `&` and `|` are used in such a way that their parsing will *not* change. The simplest way I know of marking such locations is by changing its symbol in the AST (to `_precedence_and` and `_precedence_or`).

We want, however, for the obfuscated symbol to be the one in the *bad* locations so it may be deprecated. As such, at the very top-most level (`julia-parse`), we must walk through the entire AST and swap the symbols back.