Fix #6907.
Changes the default p in norm(::SparseMatrixCSC, p) to be p=2, to be consistent with other norm functions.

At the same time, document that norm(::SparseMatrixCSC, p=2) is currently unimplemented. Throw more helpful error message.

Note: the behavior now is that norm(::SparseMatrixCSC) will throw an error by default, as opposed to silently computing the 1-norm.