Fix this to work with the new auto-hygienic macros.
Specifically:

gen_call_with_extracted_types() is declared hygienic, and uses quote
instead of manually constructing Expr's so the result will be hygienic
as it was before the new hygienic macros.

Fix it so "@which(a[i]=x)" will work.

Fix the too-complex error message not to assume @which is the caller.

Gratuitous rearrangement of code in the for fname loop.

Also tests have been added to make testall to verify all this.  The
tests failed with the new auto-hygienic macros before these fixes
were put in.