Added StackOverflow link on README.md
I think StackOverflow is a very good platform, and helps expose Julia to a broader public. Putting a link here on the README should direct users to post more questions there.

Now, Julia is not very present at StackOverflow, and users might have the wrong impression that nobody uses the language or care about it. After all, the julia-users list is full of questions and nice answers.