Make sure SIGUSR1 is unblocked for profiler
We cannot make any assumptions about the state of the signal mask.
Turns out that the USR1 signal can be blocked by parent processes and
then the profiler will not work right (gdm is an example).  Make sure
that if we are running the profiler that the USR1 signal is unblocked.