Enable tracking of repeated consecutive keystrokes
Detect and count the number of repeated, consecutive keystrokes within the keymaps.  This supersedes the REPL's own counting of consecutive returns to abort an incomplete AST entry within its on_enter done-check with a more general implementation. The on_enter functionality is very minutely different -- instead of checking for three consecutive newlines, this checks for three consecutive keystrokes that call the return callback (so, e.g., alternating between ^j and return won't trigger this).  But for nearly all use-cases this behaves the same.