REPL: Moving left and right was skipping newlines
Issue #7210 fixed movement across combining characters, such that moving left
and right by one character would skip all zero-width characters, but this also
inadvertently includes newlines. This made it impossible to navigate to the end
of a line that ended with a newline with the left and right arrows alone. And movement across multiple newlines would skip across them all at once.

This fixes that and adds simple regression tests.