prefix echo and printf with env in test/spawn.jl
This allows this test to pass using just the cut-down version of coreutils that we
currently bundle as part of msysgit, where echo and printf are shell scripts which
cannot be properly spawned. Slightly hacky, and runs into some newline differences at

run(`env echo hello world` |> file)

which results in a CRLF terminated file if you run it in MSYS2 or standalone,
but just LF if run inside Cygwin.