Faster filt!
The array pointer loads cannot be hoisted out of the loop if this code
is in the same function that does the checking and potential
modification of b and a. This change gives a 1.2-1.5x perf gain. It
also mitigates potential effects of type instability of b and a due to
coefficient normalization.