windows: installer fixes
"You have to have one of the new-fangled Windows 7 GUIDs in your
manifest to avoid PCA annoyance." [1]

So we do that now for the self-extracting wrapper, and also set
the name there that the download says "Julia" instead of "7zip".

7zS.sfx edits were done with ResEdit [2].

NOTE: modifying a PE manifest can be done once per session in ResEdit.
      Saving multiple time in a session leads to "side-by-side config
      failed" installer errors. I hope someone reads this someday and
      saves an hour.

[1] http://blogs.msdn.com/b/cjacks/archive/2009/06/18/pca-changes-for-windows-7-how-to-tell-us-you-are-not-an-installer-take-2-because-we-changed-the-rules-on-you.aspx
[2] http://www.resedit.net