Missing 'fully implemented' spec for Iterable
The section on Iterable Collections is missing a specification of which types fully and partly implement them.  For example, maximum is implemented for IntSet but not maxabs.  I don't know how to fix this because I don't know what is implemented and what is not.  Perhaps this is a moving target?