The macro (at)test_repr wasn't testing
This re-enables testing of representations of Exprs. Instead of testing the parsed form against the original, the (at)test_repr macro was comparing the parsed form against the re-parsed form. It was effectively testing the stability of `show` instead of the actual parsing.

Only two tests failed upon fixing the macro to test the re-parsed form against the explicitly written string.  One was simply due to empty lines that are not in the parsed representation.  The other is `w = (x = y) = z`, which displays simply as `w = x = y = z`.  The Expr structure of the two is indeed different, but the tested form is an invalid assignment location and of dubious utility.

Fixes #7879