Adjust the default threshold parameters for the least squares
The present tolerance is larger than necessary in the implementation
of ldiv operator which can lead to loss of accuracy. Set the default
tolerance parameter that is compatible with the conservative MATLAB estimate
for determining the numerical rank of A, i.e., tol= eps()*maximum(size(A)).

In addition, fix the implementation of the SVD based least squares,
and increase the default tolerance parameter for the Moore–Penrose
pseudoinverse which can not be determined to full machine precision
for ill-conditioned matrices.