Added forward compatibility aliases for #8175
After renaming some constants in #8175, it became harder to write the
same code to target 0.3.x and 0.4-dev. This PR potentially adds aliases,
so that when 0.3.1 gets released, less mainanance intrested package
authors can drop support for 0.3.0, and be compatible with both 0.3.1
and 0.4-dev with the same code base, without version conditionals.