Update modules.rst
Hello,

reading the Julia modules documentation I found it a bit misleading IMHO.

So I updated the summary table in a way that is showing that in any case, it is possible to overload the module functions when whole module or any of its function is brought into the current scope using "using" or "import" statements. Instead what changes is the cases "exports" are brought into scope or not and when it is possible to overload it or not.

I am not a Julia export so please let me know if I am mistaken.