checkcorrectness function for dict
This proposed unexported function, checkcorrectness, accompanies a proposed change that I just made to the documentation.  As noted in a recent discussion on the julia user's group, https://groups.google.com/forum/#!topic/julia-users/bXyLf5v3uBE, the Dict indexing structure is corrupted if a mutable key gets changed by the user.  The purpose of this function is to look for errors in a Dict in case the user inadvertently clobbers the index.