ByteVec: remove index dependency from the here/there branch.
I noticed that the clever trick that I used to avoid a branch in
the fast path for bytevec_ref when the bytes are immediate was
causing a needless dependency on the index in the branch condition
even in the case with no bounds checking. This could, of course,
prevent certain optimization transforms, so I got rid of it.