IteratorsMD: allow AbstractArray methods to be overriden
Any AbstractArray type that wants to override an indexing methods would
have formerly run into trouble:

    next{T,N}(A::SomeArrayType{T,N}, state::CartesianIndex{N})

would be more specific in the first argument, but (because
CartesianIndex{N} is less specific than CartesianIndex_$N) would be
less specific in the second argument. So there was no way to avoid
a method ambiguity warning.