Fix help for operators starting with a dot
This fixes #8903 and supersedes #8922 (cc @hayd). Basically, when you pass a function to the help system (instead of just its name), the system can go through all the modules and only returns those methods that actually extend the same function.  The way it does this, however, is a little hacky... using string concatenation to combine the function name with the module names, and then splitting it back apart again on dots to get the functions within each module.  This simply improves the splitting logic with a regex to handle the leading dot.

(cherry picked from commit 56713813556d207c44f3495ad3120cf671f4a929)
Ref: #9109