merge 2 versions (for GMP 5 & 6) of rand(::UnitRange{BigInt})
As a consequence, using dynamically a version of GMP which
does not match the compile time version is no longer an error
(this patch makes this a warming).
The code is shorter, but we break the MPZ API boundary; this also
allows to put allocating pointer_to_array out of the loop, and get
slightly better performances.