re-enable the bell ("beep")
Feedback is important. When I hit `tab` I shouldn't be stuck wondering
if the system isn't responding, or if there are just no completions.

I also agree with those that hate the audio bell but what to do with
'\a' is a terminal setting, and most terminals can be configured to
give a visual bell (a brief flash), or disable completely. So this is
not something that each shell/program needs to also have configuration
for.

How to configure the bell for assorted terminals:

  OSX: http://superuser.com/questions/163994/how-can-i-get-the-mac-terminal-to-not-beep
  Gnome Terminal: Edit -> Profile Preferences -> Terminal Bell
  KDE Konsole: Settings -> Configure Notifications
  XTerm: add the following to .Xdefaults/.Xresources: xterm*visualBell:true
  URxvt: add the following to .Xdefaults/.Xresources: URxvt*visualBell:true
  XWindows: disable across all programs by runnings: xset b off