handle GenSym objects in interpreted code, correct issues with julia-syntax.scm handling of (jlgensym) objects, and use of (make-jlgensym) where possible, disable a test that this (temporarily) broke
julia-syntax optimistically assumes many variables will be jlgensym-compatible. it is the responsiblity of branch assignment locations (such as if blocks) to handle this case appropriately, and only emit one assignment to the (jlgensym? dest) variable.

the type-inference information for GenSym variables now may be just the count of the number of GenSym variable slots present in the function