Change `jldownload` to use the caching server for all queries.
Get rid of `$JLDOWLOAD_PREFER_MIRROR` since the caching server transparently caches now, and falls back to redirecting to the original URL in the case of a cache miss.
Note that we keep a fallback to directly asking for the normal URL in the unlikely event of the cache service completely failing.