In inference, make string constants module local
During inference, we can end up in a situatio where the stack looks like

COMPILE B
CALL B
COMPILE A
CALL A

If A and B are methods of the same functions, it is likely that they share
string constants, but in MCJIT these will be allocated in the module for A,
so when B references them, the call to B tries to materialze the modules for
both A and B, leading to an assertion failure, because the compile for A is
not finished yet. Work around this by disabling string constant pooling in
inference. The proper fix is to write an LLVM pass that records the location
of global constants before the module is emitted and merges them there.