Dup FD when converting an IO object to a FILE*
@JeffBezanson and I decided that was the best way to go about handling the issue that
fdopen wants to take ownership of the file descriptor. In this way the two objects
are independent and everything should work in the general case. Also add a close method
to Libc.FILE, which can now always be safely called (assuming of course that it isn't
closed by the C function it gets passed to).

Finally, move RawFD and dup to Libc, since they're part of the C library anyway.