Fix Pkg.update() for empty ~/.julia/v0.4
By my own clever use of the `unison` package [0], I ran into a
situation where ~/.julia/v0.4 existed but did not contain a METADATA
directory.  If ~/.julia/v0.4 exists but ~/.julia/v0.4/METADATA does
not, Pkg.update() fails with an obscure error message.  This patch
fixes this.

In addition, I noticed that if the git clone of METADATA fails, the
exception handling removes everything in ~/.julia/v0.4.  If the user
has packages there already, this can be overkill (and lead to data
loss if there are local changes), so I changed it to remove only
everything in ~/.julia/v0.4/METADATA when METADATA fails to clone.

[0] http://www.cis.upenn.edu/~bcpierce/unison/